requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-chroma
chromadb
streamlit
fastapi
uvicorn
python-dotenv
openai
tiktoken
pydantic
requests
langchain-groq
